LOOK HERE FOR H1 ALIGNMENT FOR H2,H3
This is the commit that should be referred to for code in case i change
my mind about where to put the flower pictures. So far, I am still
wanting to put them to the right of the h2, h3 titles. So after this i
am removing the padding that lets the h2, h3 line up with h1.